requirements.txt
bcrypt==4.3.0
blinker==1.8.2
click==8.1.8
colorama==0.4.6
exceptiongroup==1.3.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.1
importlib_metadata==8.5.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
packaging==25.0
pluggy==1.5.0
psutil==7.0.0
pytest==8.3.5
python-dotenv==1.0.1
SQLAlchemy==2.0.41
tomli==2.2.1
typing_extensions==4.13.2
Werkzeug==3.0.6
WTForms==3.1.2
WTForms-SQLAlchemy==0.4.1
zipp==3.20.2
